cleanup server
