+++ src/server/logic/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod init;
+pub(crate) mod movement;
