+++ src/server/mod.rs
@@ -1,4 +1,2 @@
-pub mod connection;
-pub mod init;
-pub mod movement;
-pub(crate) mod update_server_system;
+pub(crate) mod logic;
+pub(crate) mod network;
